requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
confusable-homoglyphs==3.2.0
distlib==0.3.1
Django==3.1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-filter==2.4.0
django-registration==3.1.1
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
install==1.3.4
mysqlclient==2.0.1
pbr==5.5.1
Pillow==8.0.1
Pygments==2.7.3
PySocks==1.7.1
pytz==2020.1
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.0
stevedore==3.3.0
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
